Add directive visitors support
Although functional, needs some refactoring to get rid of gross code
duplication.

Adds support for registering `:directive-visitors` when compiling the
schema. This should be a map, with directive names as keys and a
function to handle that specific directive. This function will be called
multiple times, depending on where the directive is used. The visitor
will receive a map with the following keys:

- `:category`: the specific type of visiting that is being done:
`#{:object, :field-definition, :enum, :enum-value, :scalar,
:argument-definition, :schema, :interface, :union, :input-object,
:input-field-definition}`
- `:execution-context`: An execution context as it is used to execute
the query.
- `:resolver`: The original resolver that is used to resolve an actual
value, can be used in the visitor to decorate or transform values.